Refactored legacy v1.0 project structure + fixed warnings

+++ Sirius/src/move.cpp
@@ -6,7 +6,7 @@ Move::Move(int src, int dst, MoveType type)
 {
 	assert(src >= 0 && src < 64 && "Src pos is out of range");
 	assert(dst >= 0 && dst < 64 && "Dst pos is out of range");
-	m_Data = src | (dst << 6) | static_cast<int>(type);
+	m_Data = static_cast<uint16_t>(src | (dst << 6) | static_cast<int>(type));
 }
 
 Move::Move(int src, int dst, MoveType type, Promotion promotion)
@@ -14,5 +14,5 @@ Move::Move(int src, int dst, MoveType type, Promotion promotion)
 {
 	assert(src >= 0 && src < 64 && "Src pos is out of range");
 	assert(dst >= 0 && dst < 64 && "Dst pos is out of range");
-	m_Data = src | (dst << 6) | static_cast<int>(type) | static_cast<int>(promotion);
+	m_Data = static_cast<uint16_t>(src | (dst << 6) | static_cast<int>(type) | static_cast<int>(promotion));
 }
